backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
docling==2.65.0
openai==2.13.0
python-multipart==0.0.9
python-dotenv==1.0.1

